More pieces are in place.
