+++ zingo-cli/Cargo.toml
@@ -4,11 +4,13 @@ version = "0.2.0"
 edition = "2024"
 
 [features]
-regtest = ["zingolib/testutils", "zingo_common_components/for_test"]
+regtest = ["zingolib/testutils", "zingo_common_components/for_test", "zcash_local_net"]
 
 [dependencies]
 zingolib = { workspace = true }
 pepper-sync = { workspace = true }
+
+zcash_local_net = { git = "https://github.com/zingolabs/infrastructure.git", rev = "5553bc50339bb13e208688387be8496ae59a7d5c", optional = true }
 
 zcash_protocol = { workspace = true }
